Complete Create,Edit & List Action For TenderEvaluationCommittee

+++ BGB.Data/Entities/Wbpm/ContractorCompanyPayment.cs
@@ -1,11 +1,11 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System;
 using System.ComponentModel.DataAnnotations.Schema;
+
 using BGB.Data.Entities.Base;
 
 namespace BGB.Data.Entities.Wbpm
 {
-    [Table("ContractorCompanyPayment", Schema = "wbpm")]
+	[Table("ContractorCompanyPayment", Schema = "wbpm")]
     public class ContractorCompanyPayment : BaseEntity
     {
 
@@ -15,24 +15,24 @@ namespace BGB.Data.Entities.Wbpm
 
         public int ProjectWorkId { get; set; }
         public decimal EstimatedCost { get; set; }
-        public float EstimatedCostTaxPer { get; set; }
+        public double EstimatedCostTaxPer { get; set; }
         public decimal EstimatedCostTaxAmount { get; set; }
-        public float EstimatedCostVatPer { get; set; }
+        public double EstimatedCostVatPer { get; set; }
         public decimal EstimatedCostVatAmount { get; set; }
-        public float EstimatedCostCollateralPer { get; set; }
+        public double EstimatedCostCollateralPer { get; set; }
         public decimal EstimatedCostCollateralAmount { get; set; }
         public decimal EstimatedCostDeductionAmount { get; set; }
         public decimal NetEstimatedCostAmount { get; set; }
-        public float ProgressPer { get; set; }
+        public double ProgressPer { get; set; }
         public decimal ProgressAmount { get; set; }
-        public float ProgressTaxPer { get; set; }
-        public float ProgressVatPer { get; set; }
-        public float ProgressCollateralPer { get; set; }
+        public double ProgressTaxPer { get; set; }
+        public double ProgressVatPer { get; set; }
+        public double ProgressCollateralPer { get; set; }
         public decimal ProgressDeductionAmount { get; set; }
         public decimal NetProgressAmount { get; set; }
-        public float PerformanceSecurityPer { get; set; }
+        public double PerformanceSecurityPer { get; set; }
         public decimal PerformanceSecurityAmount { get; set; }
-        public float ContactorProgressPer { get; set; }
+        public double ContactorProgressPer { get; set; }
         public decimal ContactorProgressAmount { get; set; }
         public int BillPaymentNumber { get; set; }
         public decimal PreviouslyPaidAmount { get; set; }
